requirements.txt
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
dominate==2.4.0
idna==2.9
imageio==2.8.0
jsonpatch==1.25
jsonpointer==2.0
kiwisolver==1.1.0
matplotlib==3.1.3
networkx==2.4
numpy==1.18.1
Pillow==7.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
PyWavelets==1.1.1
pyzmq==18.1.1
requests==2.23.0
scikit-image==0.16.2
scipy==1.4.1
six==1.14.0
torch==1.4.0
torchfile==0.1.0
torchvision==0.5.0
tornado==6.0.3
urllib3==1.25.8
visdom==0.1.8.9
websocket-client==0.57.0